// vhome_mis/src/app/core/api-client-v2/model/managementEmployeeResponseAuth.ts
/**
 * APIFlask
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ManagementEmployeeResponseAuth { 
    ID: number;
    account: string;
    password: string;
    totp_secret?: string | null;
}