// vhome_mis/src/app/core/api-client-v2/model/managementEmployeeResponseEmployeeDetails.ts
/**
 * APIFlask
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ManagementEmployeeResponseAuth } from './managementEmployeeResponseAuth';


export interface ManagementEmployeeResponseEmployeeDetails { 
    ID: number;
    auth?: ManagementEmployeeResponseAuth | null;
    company_email: string;
    date_of_employment: string;
    deleted: boolean;
    localization: string;
    name: string;
}